fix: 빌드 타입 에러 수정
